Set default (int) to config parameters

+++ src/config/log-alarm.php
@@ -9,13 +9,13 @@ return [
     'log_type' => env('LA_LOG_TYPE', 'error'), // also possible: 'error,warning,debug'
 
     // log time frame - log time frame to listen - in minutes
-    "log_time_frame" => env('LA_LOG_TIME_FRAME', 1),
+    "log_time_frame" => (int)env('LA_LOG_TIME_FRAME', 1),
 
     // log per time frame - How many log to count per time frame until alarm trigger 
-    "log_per_time_frame" => env('LA_LOG_PER_TIME_FRAME', 5),
+    "log_per_time_frame" => (int)env('LA_LOG_PER_TIME_FRAME', 5),
 
     // delay between alarms in minutes - How many minutes to delay between alarms
-    'delay_between_alarms' => env('LA_DELAY_BETWEEN_ALARMS', 5),
+    'delay_between_alarms' => (int)env('LA_DELAY_BETWEEN_ALARMS', 5),
 
     // log listener for specific word inside log messages
     'specific_string' => env('LA_SPECIFIC_STRING', ''), // also possible: 'table lock' or 'foo' or 'bar' or leave empty '' to enable any word
